Move range proof helper functions to range_proofs.rs.
